trying to add a chart

+++ db/migrate/20180521214729_create_votes.rb
@@ -2,7 +2,7 @@ class CreateVotes < ActiveRecord::Migration[5.2]
   def change
     create_table :votes do |t|
       t.boolean :buy
-      t.integer :amount
+      t.float :amount
 
       t.timestamps
     end
